Move shared lokihttp structures to a package that docker and k8s integration tests can re-use
